test: mixins tests (spec => test to match Deno's pattern)
